refactor(skia): move damage-region logic to Visual.Damage.skia.cs partial
Relocate the branch-new damage-region members (ContributeDamageOnPaint,
TryGetPaintDamageRegion, the local-content-bounds/path helpers, the shadow
silhouette bounds helpers, OutsetForAntialiasing, and the _lastRender*/
_subtreeChangedThisFrame fields) out of Visual.skia.cs into a dedicated
Visual.Damage.skia.cs partial. Pure relocation — no logic change; the
pre-existing render walk, picture-collapsing, and analytic-shadow code stay
in Visual.skia.cs. Keeps the core file's diff vs master near the render-loop
hooks and isolates the damage feature for review.
